+++ app/src/main/java/com/ebsolutions/papertrail/financialdataproviderservice/common/DatabaseConstants.java
@@ -3,4 +3,5 @@ package com.ebsolutions.papertrail.financialdataproviderservice.common;
 public class DatabaseConstants {
   public static final String USER_TABLE = "app_user";
   public static final String HOUSEHOLD_TABLE = "household";
+  public static final String HOUSEHOLD_MEMBER_TABLE = "household_member";
 }