+++ app/src/test/java/com/ebsolutions/papertrail/financialdataproviderservice/household/HouseholdCreateAllSteps.java
@@ -8,6 +8,8 @@ import com.ebsolutions.papertrail.financialdataproviderservice.common.exception.
 import com.ebsolutions.papertrail.financialdataproviderservice.config.Constants;
 import com.ebsolutions.papertrail.financialdataproviderservice.model.ErrorResponse;
 import com.ebsolutions.papertrail.financialdataproviderservice.tooling.BaseTest;
+import com.ebsolutions.papertrail.financialdataproviderservice.util.CommonTestUtil;
+import com.ebsolutions.papertrail.financialdataproviderservice.util.HouseholdTestUtil;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.And;
@@ -26,7 +28,6 @@ import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.test.web.servlet.MvcResult;
 
-@SuppressWarnings("checkstyle:LineLength")
 @RequiredArgsConstructor
 public class HouseholdCreateAllSteps extends BaseTest {
   protected final HouseholdRepository householdRepository;
@@ -73,7 +74,7 @@ public class HouseholdCreateAllSteps extends BaseTest {
 
     Household inputHouseholdOne = Household.builder()
         .householdId(householdId)
-        .name(HouseholdTestUtil.isEmptyString(dataTable.column(1).getFirst()))
+        .name(CommonTestUtil.isEmptyString(dataTable.column(1).getFirst()))
         .build();
 
     requestContent =
@@ -112,11 +113,11 @@ public class HouseholdCreateAllSteps extends BaseTest {
     List<Household> households = objectMapper.readerForListOf(Household.class).readValue(content);
 
     Household householdOne = households.getFirst();
-    HouseholdTestUtil.assertExpectedHouseholdAgainstActualHousehold(expectedHouseholdOne,
+    HouseholdTestUtil.assertExpectedAgainstActual(expectedHouseholdOne,
         householdOne);
 
     Household householdTwo = households.getLast();
-    HouseholdTestUtil.assertExpectedHouseholdAgainstActualHousehold(expectedHouseholdTwo,
+    HouseholdTestUtil.assertExpectedAgainstActual(expectedHouseholdTwo,
         householdTwo);
   }