+++ app/src/test/java/com/ebsolutions/papertrail/financialdataproviderservice/household/HouseholdGetAllSteps.java
@@ -7,6 +7,7 @@ import com.ebsolutions.papertrail.financialdataproviderservice.common.exception.
 import com.ebsolutions.papertrail.financialdataproviderservice.config.Constants;
 import com.ebsolutions.papertrail.financialdataproviderservice.model.ErrorResponse;
 import com.ebsolutions.papertrail.financialdataproviderservice.tooling.BaseTest;
+import com.ebsolutions.papertrail.financialdataproviderservice.util.HouseholdTestUtil;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Then;
@@ -75,16 +76,16 @@ public class HouseholdGetAllSteps extends BaseTest {
     List<Household> households = objectMapper.readerForListOf(Household.class).readValue(content);
 
     Household householdOne = households.getFirst();
-    HouseholdTestUtil.assertExpectedHouseholdAgainstActualHousehold(expectedHouseholdOne,
+    HouseholdTestUtil.assertExpectedAgainstActual(expectedHouseholdOne,
         householdOne);
 
     Household householdTwo = households.getLast();
-    HouseholdTestUtil.assertExpectedHouseholdAgainstActualHousehold(expectedHouseholdTwo,
+    HouseholdTestUtil.assertExpectedAgainstActual(expectedHouseholdTwo,
         householdTwo);
   }
 
   @Then("the correct empty households response is returned")
-  public void theCorrectEmptyHouseholdsResponseIsReturned() throws UnsupportedEncodingException {
+  public void theCorrectEmptyHouseholdsResponseIsReturned() {
     MockHttpServletResponse mockHttpServletResponse = result.getResponse();
     Assertions.assertEquals(HttpStatus.NO_CONTENT.value(), mockHttpServletResponse.getStatus());
   }